Modify data  type of protocol error

+++ invar-core/src/main/java/invar/lang/TokenParser.java
@@ -272,7 +272,7 @@ public class TokenParser {
     }
 
     private void AddProtocErrField(TypeStruct t, InvarContext ctx) throws Exception {
-        InvarType typeBasic = ctx.findBuildInType(TypeID.UINT16);
+        InvarType typeBasic = ctx.findBuildInType(TypeID.INT32);
         InvarField field = makeAutoAddField(t, typeBasic, "protocError", "Protocol error code", false);
         field.setDefault("0");
         t.addField(field);
